# docs/swagger.yaml
basePath: /api/
definitions:
...
  domain.CreateProductMasukRequest:
    properties:
      name:
        type: string
      product_id:
        type: string
      qty:
        type: string
      supplier_id:
        type: string
    required:
    - name
    - product_id
    - qty
    - supplier_id
    type: object
  domain.CreateProductRequest:
    properties:
      category_id:
        type: string
      image:
        type: string
      name:
        type: string
      qty:
        type: string
    required:
    - category_id
    - image
    - name
    - qty
    type: object
  domain.CreateSaleRequest:
    properties:
      alamat:
        maxLength: 1000
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        minLength: 12
        type: string
    required:
    - alamat
    - email
    - name
    - phone
    type: object
...
  domain.ErrorMessage:
    properties:
      error:
        type: boolean
      message:
        type: string
      statusCode:
        type: integer
    type: object
...
  domain.Response:
    properties:
      data: {}
      message:
        type: string
      statusCode:
        type: integer
    type: object
...
  domain.UpdateProductMasukRequest:
    properties:
      id:
        type: string
      name:
        type: string
      product_id:
        type: string
      qty:
        type: string
      supplier_id:
        type: string
    required:
    - name
    - product_id
    - qty
    - supplier_id
    type: object
  domain.UpdateProductRequest:
    properties:
      category_id:
        type: string
      id:
        type: string
      image:
        type: string
      name:
        type: string
      qty:
        type: string
    required:
    - category_id
    - image
    - name
    - qty
    type: object
  domain.UpdateSaleRequest:
    properties:
      alamat:
        maxLength: 1000
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        minLength: 12
        type: string
    required:
    - alamat
    - email
    - id
    - name
    - phone
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: REST API for FiberInventory
  title: FiberInventory
  version: "1.0"
paths:
...
  /api/productmasuk/{id}:
    delete:
      description: Delete a specific ProductMasuk
      parameters:
      - description: ProductMasuk ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Delete ProductMasuk
      tags:
      - ProductMasuk
    get:
      description: Retrieve the result for a specific ProductMasuk
      parameters:
      - description: ProductMasuk ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Get ProductMasuk result
      tags:
      - ProductMasuk
    put:
      description: Update a specific ProductMasuk
      parameters:
      - description: ProductMasuk ID
        in: path
        name: id
        required: true
        type: string
      - description: ProductMasuk Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateProductMasukRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Update ProductMasuk
      tags:
      - ProductMasuk
...
  /product:
    get:
      description: Retrieve the results for each product
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Get product results
      tags:
      - Product
  /product/{id}:
    delete:
      description: Delete a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - Product
    get:
      description: Retrieve the result for a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Get product result
      tags:
      - Product
    put:
      description: Update a specific product
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - Product
  /product/create:
    post:
      consumes:
      - application/json
      description: Create Product
      parameters:
      - description: Product information
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/domain.CreateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: createProduct to the application
      tags:
      - product
  /product/hello:
    get:
      description: Return a greeting message to the product
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Greet the Product
      tags:
      - Product
...
  /productmasuk:
    get:
      description: Retrieve the results for all ProductMasuk
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Get ProductMasuk results
      tags:
      - ProductMasuk
  /productmasuk/create:
    post:
      consumes:
      - application/json
      description: Create ProductMasuk
      parameters:
      - description: ProductMasuk information
        in: body
        name: productmasuk
        required: true
        schema:
          $ref: '#/definitions/domain.CreateProductMasukRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Create a new ProductMasuk
      tags:
      - ProductMasuk
...
  /sale:
    get:
      description: Retrieve the results for each Sale
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Get Sale results
      tags:
      - Sale
  /sale/{id}:
    delete:
      description: Delete a specific Sale
      parameters:
      - description: Sale ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Delete Sale
      tags:
      - Sale
    get:
      description: Retrieve the result for a specific Sale
      parameters:
      - description: Sale ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Get Sale result
      tags:
      - Sale
    put:
      description: Update a specific Sale
      parameters:
      - description: Sale ID
        in: path
        name: id
        required: true
        type: string
      - description: Sale Data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateSaleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Update Sale
      tags:
      - Sale
  /sale/create:
    post:
      consumes:
      - application/json
      description: Create a new Sale
      parameters:
      - description: Sale information
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateSaleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.ErrorMessage'
      security:
      - BearerAuth: []
      summary: Create a Sale
      tags:
      - Sale
...
securityDefinitions:
  BearerAuth:
    in: Header
    name: Authorization
    type: apiKey
swagger: "2.0"